McMurchie AND DAVIDSON (MD) METHOD IMPLEMENT
use red-black tree store and search F(m, w).
but the time compute ERI more than don't use it.

and the memory comsumption increase quickly.

+++ redblack.h
@@ -29,7 +29,7 @@
 #ifndef _REDBLACK_H
 //#include "common.h"
 typedef struct {
-    unsigned int w;
+    double w;
     double value;
 }ITEM;
